backend/requirements.txt
# FastAPI and ASGI
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
alembic==1.12.1
aiosqlite==0.19.0
asyncpg==0.29.0  # For PostgreSQL support

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# HTTP Clients
httpx==0.25.2
aiohttp==3.9.1

# Background Tasks
apscheduler==3.10.4
celery==5.3.4
redis==5.0.1

# Data Validation
pydantic==2.5.0
pydantic-settings==2.1.0

# Crypto & Math
cryptography>=42.0.0

# External API Clients
openai==1.3.7
websockets==12.0

# Utilities
python-dateutil==2.8.2
pytz==2023.3
requests==2.31.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0
httpx==0.25.2

# Development
black==23.11.0
isort==5.12.0
flake8==6.1.0
